requirements.txt
aiogram
SQLAlchemy
asyncpg
greenlet
python-docx
python-dotenv
pydantic
aiofiles
